Add interactivity to post-install

+++ post-install.sh
@@ -10,21 +10,32 @@
 source ./lib_sh/echos.sh
 source ./lib_sh/requirers.sh
 
+# Set desired terminal theme name
+# https://github.com/lysyi3m/macos-terminal-themes.git
+TERMINAL_THEME="FrontEndDelight"
+
+
 ###############################################################################
 # ZSH                                                                         #
 ###############################################################################
 
 bot "finishing ZSH installation"
 
-running "installing PowerLine fonts"
-./Sites/_Tools/powerline-fonts/install.sh
+read -r -p "Do you want to install Powerline fonts? [Y|n] " response
+if [[ $response =~ ^(y|yes|Y) ]];then
+    running "installing PowerLine fonts"
+    ./Sites/_Tools/powerline-fonts/install.sh
+fi
 ok
 
-running "activating zsh-completions"
-# make sure the path has the correct permissions to avoid insecrure directories warning:
-# https://stackoverflow.com/questions/13762280/zsh-compinit-insecure-directories
-sudo chmod -R 755 /usr/local/share
-rm -f ~/.zcompdump; compinit
+read -r -p "Do you want to activate zsh-completions? [Y|n] " response
+if [[ $response =~ ^(y|yes|Y) ]];then
+    running "activating zsh-completions"
+    # make sure the path has the correct permissions to avoid insecrure directories warning:
+    # https://stackoverflow.com/questions/13762280/zsh-compinit-insecure-directories
+    sudo chmod -R 755 /usr/local/share
+    rm -f ~/.zcompdump; compinit
+fi
 ok
 
 running "making sure ZSH is up to date"
@@ -47,3 +58,25 @@ npm config set save-exact true
 ok
 
 
+###############################################################################
+# TERMINAL                                                                    #
+###############################################################################
+
+sleep 1
+bot "To finish the Terminal customisation..."
+sleep 2
+running "Open Terminal > Prefernces (Cmd+,)"
+sleep 2
+running "Select the ${TERMINAL_THEME} theme"
+sleep 2
+running "Set a PowerLine font family (Ex: Meslo LG S DZ Regular Powerline"
+sleep 4
+read -r -p "Restart terminal? [Y|n] " response
+if [[ $response =~ ^(y|yes|Y) ]];then
+    action "killing terminal..."
+    sleep 2
+    killall "Terminal" &> /dev/null
+else
+    ok "terminal left open";
+    info "Remember that changes will only take place once you restart the terminal or in new windows"
+fi
